move metadata to layout
