+++ template/web/app/buy-me-coffee/page.tsx
@@ -1,19 +1,38 @@
-import { generateMetadata } from '@/utils/generateMetadata';
-import BuyMeCoffeePage from '.';
+'use client';
 
-export const metadata = generateMetadata({
-  title: 'Build Onchain Apps Toolkit',
-  description: 'Build Onchain Applications with the best consumer experience in a few minutes.',
-  images: 'themes.png',
-  pathname: 'buy-me-coffee',
-});
+import { useEffect, useState } from 'react';
+import Banner from '@/components/layout/banner/banner';
+import Footer from '@/components/layout/footer/Footer';
+import Header from '@/components/layout/header/Header';
+import Main from '@/components/layout/Main';
+import BuyMeCoffeeContractDemo from './_components/ContractDemo';
+import Guide from './_components/Guide';
 
 /**
- * Server component, which imports the BuyMeCoffeePage component (client component that has 'use client' in it)
- * https://nextjs.org/docs/app/building-your-application/routing/pages-and-layouts
- * https://nextjs.org/docs/pages/building-your-application/upgrading/app-router-migration#step-4-migrating-pages
- * https://nextjs.org/docs/app/building-your-application/rendering/client-components
+ * Use the page component to wrap the components
+ * that you want to render on the page.
  */
-export default function Page() {
-  return <BuyMeCoffeePage />;
+export default function BuyMeCoffeePage() {
+  const [isMounted, setIsMounted] = useState(false);
+
+  useEffect(() => {
+    setIsMounted(true);
+  }, []);
+
+  //  Fix hydration issues
+  if (!isMounted) {
+    return null;
+  }
+
+  return (
+    <>
+      <Header />
+      <Main>
+        <Banner pageName="Buy Me A Coffee" pageUrl="buy-me-coffee" />
+        <BuyMeCoffeeContractDemo />
+        <Guide />
+      </Main>
+      <Footer />
+    </>
+  );
 }
