+++ template/web/app/paymaster-bundler/layout.tsx
@@ -1,0 +1,12 @@
+import { generateMetadata } from '@/utils/generateMetadata';
+
+export const metadata = generateMetadata({
+  title: 'Build Onchain Apps Toolkit',
+  description: 'Build Onchain Applications with the best consumer experience in a few minutes.',
+  images: 'themes.png',
+  pathname: 'paymaster-bundler',
+});
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return children;
+}
